Changing the reference schema for Netlify CMS (#9)
This pull request changes the schema for adding references.

Previously we added references by creating a list of Markdown strings. This was difficult to write and difficult to parse.

```yaml
references:
 - "[Adaptive Subgradient Methods for Online Learning and Stochastic Optimization](http://jmlr.org/papers/v12/duchi11a.html)"
 - "[An overview of gradient descent optimization algorithms - Sebiastian Ruder](http://sebastianruder.com/optimizing-gradient-descent/)"
```

Now we are using the Netlify CMS, it makes a lot more sense to have the URL and title text as separate fields. This is part of a broader shift from an text editor-based editing workflow to a CMS-based editing workflow.

In the YAML front matter, this is what the new reference style looks like:

```yaml
references:
- link_title: Adaptive Subgradient Methods for Online Learning and Stochastic Optimization
  link_url: http://jmlr.org/papers/v12/duchi11a.html
- link_title: An overview of gradient descent optimization algorithms - Sebiastian
    Ruder
  link_url: http://sebastianruder.com/optimizing-gradient-descent/
```

Additionally, we now list the website's domain name in the hyperlink title.