Bringing back the Dockerfile
It turns out that it's a pain in the butt to
try and test *locally* on my Mac without Docker, so
the Dockerfile is coming back for others that would
like an out-of-the-box way to run the site locally.

However, we won't be using this Dockerfile for actual deploys,
because Netlify just does that with so much less fuss.