Add a sample term to test the possible Pandoc outputs
We can't easily use Tachyons class names to style the
HTML output of Pandoc, so this file is to generate
and test all the different types of HTML that Pandoc
might throw at us.

We'll need to use the SCSS `@extend` feature to style
regular HTML elements with all of the Tachyons classes
we want to use.